+++ livre/core/data/VolumeInformation.h
@@ -87,13 +87,23 @@ struct VolumeInformation
     /** The normalized size of the volume. */
     Vector3f worldSize;
 
-    /** Real-world size of the volume in meters. */
-    Boxf boundingBox;
+    /** The transformation matrix between data space and livre space */
+    Matrix4f dataToLivreTransform;
+
+    /** The resolution in "data units"/voxel. Only valid for some datasources,
+     * ( -1, -1, -1) otherwise.
+     */
+    Vector3f resolution;
 
     /**
      * The world space per voxel. Assumes the volume is isotropic.
      */
     float worldSpacePerVoxel;
+
+    /**
+     * The ratio between a meter and the unit used by the data.
+     */
+    float meterToDataUnitRatio;
 
     /**
       * Root node info