+++ livre/eq/Config.cpp
@@ -45,6 +45,7 @@
 #include <eq/eq.h>
 
 #include <deque>
+#include <functional>
 
 namespace livre
 {
@@ -91,7 +92,6 @@ public:
         , defaultLatency( 0 )
         , currentCanvas( 0 )
         , eventMapper( EventHandlerFactoryPtr( new EqEventHandlerFactory ))
-        , volumeBBox( Boxf::makeUnitBox( ))
         , redraw( true )
         , dataFrameRange( INVALID_FRAME_RANGE )
     {}
@@ -157,7 +157,6 @@ public:
     eq::Canvas* currentCanvas;
     EventMapper eventMapper;
     FrameData framedata;
-    Boxf volumeBBox;
 #ifdef LIVRE_USE_ZEROEQ
     std::unique_ptr< zeroeq::Communicator > communicator;
 #endif
@@ -242,15 +241,16 @@ void Config::resetCamera()
         getApplicationParameters().cameraPosition );
     _impl->framedata.getCameraSettings().setCameraLookAt(
         getApplicationParameters().cameraLookAt );
-#ifdef LIVRE_USE_ZEROEQ
-    _impl->communicator->publishCamera();
-#endif
 }
 
 bool Config::init( const int argc LB_UNUSED, char** argv LB_UNUSED )
 {
 #ifdef LIVRE_USE_ZEROEQ
     _impl->communicator.reset( new zeroeq::Communicator( *this, argc, argv ));
+
+    _impl->framedata.getCameraSettings().registerNotifyChanged(
+                std::bind( &zeroeq::Communicator::publishCamera,
+                           _impl->communicator.get(), std::placeholders::_1 ));
 #endif
 
     resetCamera();
@@ -483,10 +483,6 @@ bool Config::handleEvent( const eq::ConfigEvent* event )
 
     if( hasEvent )
     {
-#ifdef LIVRE_USE_ZEROEQ
-        if( _impl->framedata.getCameraSettings() != cameraSettings )
-            _impl->communicator->publishCamera();
-#endif
         _impl->redraw = true;
         return true;
     }
@@ -499,10 +495,6 @@ bool Config::handleEvent( eq::EventICommand command )
 {
     switch( command.getEventType( ))
     {
-    case VOLUME_BOUNDING_BOX:
-        _impl->volumeBBox = command.read< Boxf >();
-        return false;
-
     case VOLUME_FRAME_RANGE:
     {
         _impl->dataFrameRange = command.read< Vector2ui >();