+++ livre/eq/Event.h
@@ -28,8 +28,7 @@ namespace livre
 
 enum ConfigEventType
 {
-    VOLUME_BOUNDING_BOX = eq::Event::USER,
-    GRAB_IMAGE,
+    GRAB_IMAGE = eq::Event::USER,
     VOLUME_FRAME_RANGE,
     REDRAW,
     HISTOGRAM_DATA