+++ doc/Changelog.md
@@ -3,6 +3,15 @@ Changelog {#Changelog}
 
 # master {#master}
 
+* [#313](https://github.com/BlueBrain/Livre/pull/313):
+  Fixed the camera synchronization between applications.
+  Replaced the data bounding box by the data to Livre spaces 
+  transform matrix computed by the datasource for camera 
+  synchronization. 
+* Added meterToDataUnitRatio parameter 
+  computed by the datasource. 
+* Fixed resolution printing. 
+
 # Release 0.6 (2016-06-30)
 
 * [#315](https://github.com/BlueBrain/Livre/pull/315):